Update/dataproc persistent logging (#259)
* stashing wip before flight

* handle redirects to look at executor stderr stdout form spark ui

* fix spark logging redirect

* removing tfstate

* add alway scan user dir config in prep for hadoop 2.10

* terraform fmt

* enable component gateway on history server cluster yaml

* add component gateway to ephemeral cluster template

* variablize init paths

+++ examples/dataproc-persistent-history-server/init_actions/disable_history_servers.sh
@@ -1,7 +1,15 @@
 #!/bin/bash
 set -euxo pipefail
 role="$(/usr/share/google/get_metadata_value attributes/dataproc-role)"
-if [[ "${role}" == 'Master' ]]; then
-  systemctl stop hadoop-mapreduce-historyserver
+SPARK_HISTORY_SERVER_PATH=/etc/init.d/spark-history-server
+MR_HISTORY_SERVER_PATH=/etc/init.d/hadoop-mapreduce-historyserver
+
+if [[ "${role}" == "Master" && -f ${SPARK_HISTORY_SERVER_PATH} ]]; then
+
   systemctl stop spark-history-server
+fi
+
+if [[ "${role}" == "Master" && -f ${MR_HISTORY_SERVER_PATH} ]]; then
+
+  systemctl stop hadoop-mapreduce-historyserver
 fi
